+++ src/components/EditableHeading/EditableHeading.jsx
@@ -8,7 +8,6 @@ import EditableInput from "../EditableInput/EditableInput";
 import usePrevious from "../../hooks/usePrevious";
 import { InputType } from "../EditableInput/EditableInputConstants";
 import "./EditableHeading.scss";
-import { FontWeights } from "../Heading/HeadingConstants";
 
 const EditableHeading = props => {
   const {
@@ -128,8 +127,7 @@ const EditableHeading = props => {
       ellipsisMaxLines: props.ellipsisMaxLines,
       nonEllipsisTooltip: props.tooltip,
       size: props.size,
-      highlightTerm,
-      weight: props.weight
+      highlightTerm
     };
   };
   const renderContentComponent = () => {
@@ -146,13 +144,7 @@ const EditableHeading = props => {
     const inputType = props.inputType || textAreaType;
     return {
       value: valueState,
-      className: cx(
-        `editable-heading-input`,
-        `element-type-${props.type}`,
-        `font-weight-${props.weight}`,
-        `size-${props.size}`,
-        inputClassName
-      ),
+      className: cx(`editable-heading-input`, `element-type-${props.type}`, `size-${props.size}`, inputClassName),
       isValidValue: props.isValidValue,
       onChange: props.onChange,
       onKeyDown: props.onKeyDown,
@@ -174,8 +166,7 @@ const EditableHeading = props => {
       onIgnoreBlurEvent: onIgnoreBlurEventCallback,
       onError: onInputErrorCallback,
       onSuccess: onInputSuccessCallback,
-      ariaLabel: props.inputAriaLabel,
-      weight: props.weight
+      ariaLabel: props.inputAriaLabel
     };
   };
 
@@ -207,7 +198,6 @@ const EditableHeading = props => {
 
 EditableHeading.types = Heading.types;
 EditableHeading.sizes = Heading.sizes;
-EditableHeading.weights = Heading.weights;
 
 EditableHeading.propTypes = {
   /**
@@ -262,9 +252,7 @@ EditableHeading.propTypes = {
   onCancelEditing: PropTypes.func,
   /** Callback (with current value) when clicked on element that matches ignoreBlurClass */
   onIgnoreBlurEvent: PropTypes.func,
-  insetFocus: PropTypes.bool,
-  /** font weight of the editable heading */
-  weight: PropTypes.oneOf([EditableHeading.weights.light, EditableHeading.weights.normal, EditableHeading.weights.bold])
+  insetFocus: PropTypes.bool
 };
 EditableHeading.defaultProps = {
   className: "",
@@ -288,8 +276,7 @@ EditableHeading.defaultProps = {
   dataTestId: "",
   inputClassName: "",
   insetFocus: false,
-  maxLength: undefined,
-  weight: FontWeights.normal
+  maxLength: undefined
 };
 
 export default EditableHeading;