+++ src/constants/index.js
@@ -1,5 +1,3 @@
 export * from "./dialog";
-export * from "./keyCodes";
+export * from "./KeyCodes";
 export * from "./sizes";
-export * from "./positions";
-export * from "./common";